Fix n_measurements type to int

+++ main.cpp
@@ -7,7 +7,7 @@ constexpr float uncertainty = 1;       // How uncertain we are about that belief
 constexpr float process_noise = 0.01;  // "Model error" – how much we think the position could change unpredictably between steps
 constexpr float measurement_noise = 0.5; // How noisy we believe our sensor is
 constexpr float true_position = 5.0;
-constexpr float n_measurements = 20;
+constexpr int n_measurements = 20;
 constexpr bool verbose = true;
